Extract timing window UI module
