+++ dominant_control/ui/__init__.py
@@ -1,0 +1,5 @@
+"""UI components for Dominant Control."""
+
+from .timing_window import GlobalTimingWindow
+
+__all__ = ["GlobalTimingWindow"]
